+++ asgardpy/data/dl4.py
@@ -10,6 +10,7 @@ from gammapy.estimators import FluxPointsEstimator
 from gammapy.modeling import Fit
 
 from asgardpy.base import AnalysisStepBase, BaseConfig, EnergyRangeConfig
+from asgardpy.data.target import get_chi2_pval
 
 __all__ = [
     "FitAnalysisStep",
@@ -61,7 +62,21 @@ class FitAnalysisStep(AnalysisStepBase):
         final_dataset = self._set_datasets()
         self.fit_result = self.fit.run(datasets=final_dataset)
 
+        self.instrument_spectral_info["total_stat"] = self.fit_result.total_stat
+
+        chi2_, pval_ = get_chi2_pval(
+            self.instrument_spectral_info["total_stat"], self.instrument_spectral_info["DoF"]
+        )
+        self.instrument_spectral_info["chi2"] = chi2_
+        self.instrument_spectral_info["p-value"] = pval_
+
+        stat_message = (
+            f'The Chi2/dof value of the Fit is {chi2_:.2f}/{self.instrument_spectral_info["DoF"]}'
+        )
+        stat_message += f"\nand the p-value is {pval_:.3e}"
+
         self.log.info(self.fit_result)
+        self.log.info(stat_message)
 
     def _setup_fit(self):
         """
@@ -126,7 +141,7 @@ class FluxPointsAnalysisStep(AnalysisStepBase):
             n_jobs=self.config.general.n_jobs,
             parallel_backend=self.config.general.parallel_backend,
             reoptimize=self.config.flux_points_params.reoptimize,
-            **fpe_settings
+            **fpe_settings,
         )
 
     def _sort_datasets_info(self):